requirements.txt
streamlit==1.31.0
opencv-python==4.8.1.78
numpy==1.26.0
Pillow==10.1.0 